+++ json_serializable/lib/src/type_helpers/iterable_helper.dart
@@ -47,14 +47,6 @@ class IterableHelper extends TypeHelper<TypeHelperContextWithConfig> {
     // If they are not equal, then we to write out the substitution.
     if (subField != closureArg) {
       final lambda = LambdaResult.process(subField, closureArg);
-      if (context.config.useWrappers && isList) {
-        var method = '\$wrapList';
-        if (contextNullable) {
-          method = '${method}HandleNull';
-        }
-
-        return '$method<$itemType>($expression, $lambda)';
-      }
 
       expression = '$expression$optionalQuestion.map($lambda)';