+++ json_annotation/CHANGELOG.md
@@ -1,5 +1,7 @@
-## 2.2.1
+## 3.0.0
 
+- **BREAKING** Removed `JsonSerializable.useWrappers` and associated
+  `$`-prefixed helpers.
 - Require at least Dart `2.2.0`.
 
 ## 2.2.0